Remove ChIPS support
Remove the chips backend and testing. This removes a potentially
useful technique for mocking the plotting backend, but we can always
find it from the git history.

The configuration files are changed to remove chips as an option,
and to remove the chips-specific configuration options.

There is no change to the documentation in this commit.