Testing: add isinstance asserts to plot tests
Add several checks to ensure that the plot object returned by a
get_xxx_plot() call has the expected class. These were not made
before because the mocking done to support chips testing meant that
the tests would fail (thanks to the mocking). As this has now been
removed add back the instance checks.

Note that these tests don't really add that much, so they are not
essential.