Fix resample_data with error bars surrounding -1
The loop has been re-written so that it doesn't use -1 as a terminal
value.

I did try changing the "do I chose low or high value" flag value
to a boolean (suing randint), but that result causes the tests to
fail, so I have left it as is.

We could reduce the number of operations by reflecting the sampled
normal (i.e. if we want a "low" value but it returns a positive one,
multiply by -1), but that might change the statistical properties.