test: show current behavior with param linking / out-of-band
The test_link_parameter_setting test is intended to show how #742
is resolved. At present it just tests the current (broken) behavior,
since it isn't clear at this time what should happen.

Add an explicit test of evaluating a model with linked params (to
more-closely follow #742). A test with an XSPEC table model will be
added once the fix has been made (it can't be added beforehand as
it can crash the Python process).