Enforce paramter limits when value is accessed (when linked)
The issue is how to handle the case where a linked parameter is
set to a value outside the correct range (here lmdl.c0 has a min/max
range which contains 10, but mdl.gamma does not):

    mdl.gamma.max = 5
    mdl.gamma = lmdl.c0
    lmdl.c0 = 10

We can not (without a lot of work) raise an error when setting
lmdl.c0. This commit instead will raise an error the next time that
mdl.gamma.val is retrieved. So evaluating mdl at this point should
fail, but if the next call is to

    lmdl.c0 = 5

then you can evaluate mdl (since mdl.gamma is now in the valid
range).

The tests have not been updated to reflect this change in behavior,
so are expected to fail.