Add XSPEC convolution models (classes and to the UI layer)
This commit takes code that was used in the SDS "sherpa contribute"
routines and moves them into Sherpa itself. It adds Python-level
support for the XSPEC convolution models. In the SDS version I had
explicit load_xs<convolution model> routines, following on from
the existing load_psf/table style routines, but I have decided that
this adds an extra layer of complexity that may not be needed. So
this commit emphasizes "direct" creation of these convolution
models. The load_xs<> style routines could be added on top of this
commit if needed.

The XSPEC cflux convolution model (that calculates the energy
flux of a model) can be used like:

    from sherpa.astro import ui
    ui.load_pha('3c273.pi')
    ui.subtract()
    ui.notice(0.5, 7)
    ui.set_source(ui.xsphabs.gal * ui.xscflux.cfl(ui.powlaw1d.pl))
    pl.ampl.frozen = True
    ui.fit()
    ui.conf()

You can also use them with regridded models, as shown below

    convsource = ui.xszashift.zsh(ui.box1d.box + ui.const1d.bgnd)
    zsh.redshift = 0.8
    rgrid = np.arange(0.1, 23, 0.01)
    box.xlo = 10
    box.xhi = 14
    shiftsource = convsource.regrid(rgrid[:-1], rgrid[1:])
    ui.set_source(shiftsource)

Documentation relies on the XSPEC documentation, but some additions
have been made (e.g. the need for a fixed normalisation/amplitude
parameter for the flux/lumin models).