sample_flux: error out if Xrays is not True
Since we know that sample_flux is broken when Xrays=False, explicitly
error out if the user calls this. This raise a NotImplementedError
rather than something from the Sherpa error class as it is hopefully
a temporary change. An alternative would be to raise TypeError,
but that implies this is a longer-lived change.