Improve flux calculation for PHA data sets
The aim is to address #619 by accounting for partial overlap of the
selected energy range with the energy grid.

An alternative approach would be to ignore the dataspace completely
and just create a grid to use with the model, with the approproate
end points. This is the better solution, as it allows the flux
to be calculated over "any" range (at least as long as the model
is valid for that range), but would require a slightly-different
API, in that some control over the bin size would be required
(including linear or log binning). My original thought was nbins
and scale parameters (the latter being a string to support later
expansion), but this wouldn't help with the flux-density
calculation (where you only have a single point). There is also
some complexity over the number of points to use (e.g. for the
density calculation) since some XSPEC models expect there to be
more than one bin in the energy grid. I have left this note in
as I think we want to look at moving in this direction (and now
we have some improved tests to check it works).

NOTE: the eqwidth code looks like it will suffer the same issue
      as #619 but this has not been investigated.