Add a test to show #746 (fix #760)
Issue #746 was a regression casued by #444, and fixed by reverting
some of these changes in #759. This commit adds a test adapted from
@hamogu's initial report which shows this behavior. It fails with
Sherpa prior to #759 being merged (manually checked).

The problem caused by #444 seems to be when using analysis=wavelength,
which *I hypothesize* somehow confuses the cache code (perhaps
related to the attempt to support caching the results for multiple
datasets with the same grid).