Add docs for setting up all dependencies of the source build with conda
Now that we have a conda-based distribution for all of CIAO, we can
make use of it to install all the build dependencies for a Sherpa
source build. I find that particularly convenient, because it does
not require a fortran compiler, nor any knowledge of how or where the
dependencies such as xspec where build.