Add Chi2XspecVar test for source/bgnd bins with 0 counts (#356)
Add tests to compare Chi2XdspecVar to XSPEC 12.10.1m results
for PHA files with 0 counts in source and/or background bins.

The data set is synthetic to make it easier to control things;
the only scaling is the background exposure value, and there is
no filtering or grouping. This is simplified from the original
issue #356