Remove XSPEC dependency from multi-response test
The test added in #728 to fix #717 has been changed from using the
xsapec model to powlaw1d, and commentary added on what system was
used to calculate the test values. This has the advantage of
allowing the test to run when XSPEC is not available, but the main
benefit is that the model used (power-law) is much simpler, and so
less-likely to change over time (whereas the fit parameters for
xsapec changed significantly in XSPEC 12.11.0 compared to XSPEC
12.10.1 due to internal changes in the APEC/ATOMDB handling in
XSPEC).