Update XSPEC model parameters for XSPEC 12.11.0
Updates the following parameter settings for changes made to the
model.dat file in XSPEC 12.11.0. There is no way to make these
settings depend on the XSPEC model library in use, so we chose to
use the latest values (in many cases these will be compatible
but there may be cases where the values are not valid with older
XSPEC libraries):

 - XSagnsed: Htmax parameter has changed to lie within the range
   6 to 10, and defaults to 10 (previously you could use 6 to 200
   and the default was 100).

 - XSkerrbb: rflag and lflag are now marked as alwaysfrozen
   (they have been changed to '$' parameters in model.dat)

 - XSxscat: Xpos and Rext soft and hard limits have increased;
   they were 0.95/0.95 and 115/119 and are now 0.99/0.999 and
   235/240 respectively

There is also a documentation fix: a typo in the XSkerrbb name.