XSPEC 12.11.0: add XSbwcycl model
This model uses the c_<func> XSPEC interface, which we do not
provide a direct interface to. Instead, we take advantage of the
multiple APIs that XSPEC provides for (most) of the models and
use the "C" interface (that is just <func>).

There are now several models which use the c_<func> interface
which we do not use directly, but use variants of the above. This
should be looked at at some point.