Add test cases showing #790
This adds two tests that run through essentially the same analysis:
the first one is in channel space with a PHA file but no response,
and the other is with a PHA with some sort of response and is done
usingone of channel, energy, or wavelength space. The idea is to
ensure different code paths are exercised (but #790 looks to be
in the data class not the instrument classes).