Accept masked arrays for Data XXX creation
This change implements changes to use the mask of numpy arrays
when initializing DataXXX. It warns sensibly for non-numpy objects with
a mask in the OO interface when duck-typing, but not in the session
interface, because that whitelists only specific data types anyway.
Thus, this reverses the warnings added to the docs in #698.

fixes #695