Tests: check parameter distribution used by sample_energy/photon_flux
Checks that the parameter distribution is similar to the expected
values:

 - median is similar to best fit
 - standard deviation is similar to covariance

The tests use quite large tolerances (from trial and error) as there's
no reason the values will be exact.

A check on whether the sampled values lie within the soft limits of
the parameter (with a change to gamma limits to make this check
likely to be triggered) has been added but only as comments, since
the current parameter sampler does not restrict the soft limits
(perhaps by design).