Test: ensure test_load_multi_arfsrmfs is repeatable
The test did not set up the method and statistic, and this caused
problems when the order of tests was changed (when running in
parallel) and an unwanted change to the default statistic caused
the test to fail. This commit now tries to ensure a clean session
and forces a particular optimised and statistic.