Change scales parameter of calc_energy/photon_flux (fix #799)
The scales parameter for calc_energy/photon_flux was confusing,
so this commit attempts to clear this up, as well as adding
checks on the parameter to ensure it is the correct size
(i.e. matches the number of free parameters in the model).

When correlated=True, the scales argument should be the covariance
matrix (so n by n).

When correlated=False, the scales argument can be the covariance
matrix or a 1D array, in which case it is expected to be the
error values (ie square-root of the diagonal of the covariance).

Prior to this commit the behavior depended on whether the scales
argument was a NumPy array or not (for correlated=False). When it
wasn't then a 1D array of errors could be given. However, when it
was a Numpy array then it was assumed to be 2D but it would have
to be given the square root of the covariance matrix for the
sampling to work.