tests: add checks on parameter ranges used in sample_energy/photon_flux
Since the sampler only uses the hard limits this is not a very-informative
test. An artificial soft limit is placed on the gamma parameter so we can
check that it isn't being used.

This is mainly useful to describe the current behavior (in case there's
any changes in the future).