Error out if scales has non-finite values (sample_energy/photon_flux)
Ensure that error or covariance-matrix values are finite. There are
some checks further downstream that can catch other problem cases - e.g.
when a 1D array is used for the uncorrelated case then the absolute
of the values is used.

Also check for None values: this could happen when using a parameter
limit estimated from covar/conf, and the presence of None also changes
the NumPy conversion which would result in an unfriendly error message
at the isfinite check (so checking for None first removes this case).