Error out if no free parameters and change error message
The error type and message has been changed to use more-specific
Sherpa errors than ArgumentErr.

Note that the number of free parameters is defined by the model in
the fit object and not the src argument. This is due to the way
that the fit object is used in the calc_flux routine.