Tests: tweak tolerance
When comparing the parameter distributions, nH seems to require a larger
tolerance than the other parameters since it's best-fit value is
close to the lower limit boundary (0). So bump up the tolerance for only
this parameter.

We could freeze the nH value, but I want the tests to include parameters
from multiple model components here.