sample_energy/photon_flux returns only parameters used in flux calc
Now that we can send in a subset of the model - e.g. to calculate the
unabsorbed flux for a source model defined as xsphabs * powlaw1d -
there is no need to send in, or return, parameter values for the
unused components. However, there is an ease-of-use experience,
when a user might want to say

    >>> set_source(xsphabs.gal * powlaw1d.pl)
    >>> fit()
    >>> covar()
    >>> cmat = get_covar_results().extra_output
    >>> v1 = sample_energy_flux(num=1000, scales=cmat)
    >>> v2 = sample_energy_flux(num=1000, scales=cmat, model=pl)

That is, to not have to sub-set the scales argument when a model
component is defined. This is therefore also supported.

Note that the return value only contains the relevant parameters,
so in the example above v1.shape == (1000, 1 + 3) and
v2.shape == (1000, 1 + 2), since there are 3 and 2 free parameters
in "gal * pl" and "pl" resectively (the models that are used for
the flux calculation in the two cases).

The code to implement this is a bit verbose, but has been left this
way rather than trying to simplify it a bit, to make it easier
to follow (e.g. there is some code that could be factored out but
I don't think it's worth it at this time).

Several additional checks are added, to try and make sure the inputs
make sense. In particular there is a requirement that the model used
for the flux calculation is "part of" the source expression (so
in the example above we would allow gal * pl or pl [or technically
gal, but that is physically meaningless]). There is no easy way
to perform this check (as far as I can tell) - as the 'in' method
only really works for a single model component and not composite
expressions such as 'gal * pl' - so several simple checks are used.
There are likely still cases that could trip this up, but they
are being left until test cases are added for the background case.

One result of these changes is that the fit argument has been
removed from the calc_flux and CalcFluxWorker helper
routines/classes. As these are "internal" routines I feel that
it is okay to remove them here.