Support scales parameter for *energy/photon_flux* routines (#801)
The scales argument is now correctly handled for

  get_energy_flux_hist, get_photon_flux_hist
  plot_energy_flux, plot_photon_flux

This commit also removes the **kwargs support in the get* variants,
as there is currently no use for it.

The recalc argument is expected to be a boolean, but the bool_cast
routine is used just in case (it should probably be removed but
that's for later, if ever).