sample_flux: screen output is now via the logger
Use the Sherpa logger instance rather than print to display the
summary statistics. This requires putting back a number of changes
to the log level that can hopefully be removed with future clean ups.

An example is added to sample_flux to show how to take advantage of
this functionality.