calc_sample_flux: remove un-needed id and session arguments
The session and id arfuments are used to access information that
we can get in other ways, using arguments we already send in. It
also helps separate the two regimes (the sherpa.astro.flux routines
are "below" the UI layer and so should not know about identifiers
and session objects).

As this is a low-level routine I feel okay in removing arguments
from it (especially as we don't need them!).