Allow XSPEC models to be used with regrid
Supports regridding of additive and multiplicative models. The
multiplicative case is "interesting" since we have integrate=False
but are sending in the low and high edges. We could just try and
rely on the standing integrate=False handling here, but that
relies on the left-edge of the bin, and it is not at all obvious
what the XSPEC models use (since the decision is up to each
multiplicative model).

This commit removes the ability to send kwargs to the XSPEC models
(at least for non-table models) which is okay since these functions
do not support kwargs.

It is an interesting question what to do about kwargs here, but
all the models have this (in that they'll just ignore unrecognized
options).