Fix: sample_energy/photon_flux with bkg model
Support for the model parameter had broken the handling of
background models. A small change was made to calc_energy/photon_flux
for handling of background models to be more consistent in behavior.

Tests are added for calc_energy_flux and sample_energy_flux with a
background model. The sample_energy_flux test requires changes made
in this commit to run.