Add multi-dataset anaysis to sample_energy/photon_flux
By setting the otherids argument to sample_energy_flux or
sample_photon_flux the analysis will use the combined dataset
when estimating errors. The actual calculation of the flux is
unchanged from before, it is just the errors that are different.

I've tested this and with a dataset of approximately 40, 80, 40 ks
you can see the significantly smaller errors on the parameters used
to create each realisation.