sample_energy/photon_flux always uses the fit number of free parameters
In adding support for the model parameter to sample_energy_flux and
sample-energy_flux, I made sure that users could - when explicitly
giving errors via the scales argument - give either the total number
of free parameters in the full model expression or just the number
of free parameters in the model expression. This decision meant that
I chose to always return just those parameters in the model
expression in the return value.

I have decided that this is not worth the code complexity, and that
there is a use to retaining the model parameters even for components
that aren't even used (e.g. an absorption model or, when fitting the
background, any background model components). This commit therefore
removes support for having the scales parameter being smaller than
the number of free parameters in the fit model.