Tests: remove several un-necessary tests
A number of tests looped over id = None, 1, 2, "foo". There is
no real benefit to having the 2 in this list so it was removed,
which removes a non-negligible number of tests. Technically we
could also remove the 1 (e.g. just test None and "foo"), but I
felt it useful to keep 1 in the tests as it is likely a common
argument (when not left at None). I want to leave a string version
to check that this is handled properly.