sherpa.astro.flux.sample_flux - add clip parameter and extend return value
The clip parameter controls how the list of parameter values are clipped
(previously this was fixed to using hard limits). The choice is hard or
soft limits or no clipping. The return value has an extra column at the
end which indicates if the row was clipped (1) or not (0).

There's no support in the sherpa.astro.ui to pass this information
down to this function.