Move parameter clipping to sherpa.sample class
This moves the logic for clipping the sampled arrays from the flux
code down into the sampler classes. For the moment this is done as
a separate step from the get_samples routine, but maybe it should
be part of get_samples? One reason for not changing it is to make
it clearer where the clipping is happening in the existing code.