Allow the test_astro_ui::test_psf_model1d tests to run
Reworking the tests to support pytest showed that only one of the
three models in the test_psf_model1d case failed, so this adds the
tests back in (they were previously labelled as "tests used to not
be run because of a naming conflict) and changes the test so that
the center value is checked.

It is not obvious to me whether the lorentz1d center should be 1,
4, or some other value. This just lets us check the current
behavior.