explicit notice/ignore_id error message for ids=None + doc tweaks
The error message when notice_id or ignore_id are called with id
set to None has been made explicit (it doesn't actually change user
experience). I had thought about supporting a None value - we can
easily convert it to the default id - but decided against it as
people may have got confused with None as a value for lo. This can
be revisited.

The notice and ignore methods have the bkg_id argument described.
This is technically in the wrong place (it should be in the astro
version), but it saves having to duplicate a function just to add
a small-amount of information.