Support vectors for background scale factors (when fitting)
The previous version used the sherpa.astro.background.BackgroundSumModel
to handle the scaling of the separate background models for a dataset,
when the background is fitted and not subtracted (or handled by the
WSTAT statistic).

There are two issues here:

a) the str representation of the source model (i.e. value returned by
   get_model) was only an approximation of the truth (e.g. if there
   are multiple backgrounds and they have different scales then the
   screen output would say <value> * (mdl1 .. mdln), which is wrong.
   It only affects the str output (the calculation is correct), but
   this is surprising (e.g. for people trying to work out what is
   being done)

b) it fails when the scale factor is a vector (both in generating
   the str representation and when evaluating the model), because
   the scale factor (in channel space) is multiplied by the model,
   which is evaluated using the instrument response grid.

The code now drops the BackgroundSumModel (and removes it) since
the model expression syntax is sufficient to represent the full
model (i.e. no need to add a class to do this, and this makes it
clearer what model parameters are being used).

A new method of the DataPHA class has been added, which
calculates the scale factor for each background dataset (in
a different manner to existing routines). This should be reviewed
once everything has been resolved as we may want to remove or
replace the other routines.

The sherpa/astro/ui/utils.py:_add_convolution_models has been
re-worked to make it easier to see when the various options
lead to a return value (to avoid too much indentation), and
some documentation has been added.

There are at least two issues (second is more serious):

 - the full array is included in the source expression which is
   of no use to anyone. I plan to add a model to represent
   this array for book-keeping/display purposes

 - in order to support multiplying by the channel-space array,
   the expression has changed to

     resp(source_model) + scale1 * resp(bgnd_model1) + ...

   This is correct (the outout of resp(...) is in channel
   space), but this is **ONLY VALID** if the response is
   linear. This is not the case with the jdpileup model.

   The code will warn the user - using the Sherpa logging
   infrastructure rather than the Python warnings
   infrastructure - if this condition exists, but will not
   stop the model from being evaluated.

Even with this constraint, the code does add functionality to
Sherpa (the ability to use the vector scaling values, as long as
the pileup model is not used).

There is a usability issue, but this is because the code we have
that can generate vector scale arrays (normally created by some
sort of scaled addition) doesn't work well when the background bins
are 0. This is not technically an issue with this commit: the code
is doing what it's been told to do, it's just the data isn't
set up well.