Use ArithmeticConstantModel to store the vector array for background scaling
The handling of the scale array when there's a vector in the term
relating the background to source aperture has been changed to use
the ArithmeticConstantModel rather than being left as a NumPy array,
as it better fits into the model system.

Fortunately the ArithmeticConstantModel has the desired behavior -
it always returns the stored value (in this case an array in channel
space) no-matter what the evaluation grid is. This is needed because
for our use case we always return all channels, even if the data had
been filtered (the filtered channels get a 0 value).