Add tests for background handling
Add a check of the statistic calculation as a regression test
(i.e. the values are calculated using this code, rather than
first principles). The filtered case is currently marked as
an error (I was expecting to have this problem somewhere but
was finding it hard to trigger with the existing test cases).

Try to improve checks for potential errors (to make sure you
don't get different errors depending on the code path, which has
been seen during development).