Fix handling of filtered data with backscale array
The trick is to replace a tablemodel with a model which always returns
the full data array (there doesn't appear to be a way to do this with
a tablemodel since the model gets sent a subset of the channels but
needs to return the full array, which tablemodel doesn't support).

This relies on the RSP model always returning the full channel space,
even if it only evaluates for a subset of the channel grid. I believe
I added this behavior to handle the AREASCAL case in #351 but I
haven't dug deeply into that.