Change background scaling for source regions
When subtracting the background data you say

  source - r_exposure * r_backscale * r_areascal * bgnd

where r_x is the ratio of source x to background x (this assumes a
single background dataset but it extends to multiple
datasets).

When modelling the background we have to create the model (excluding
the response term)

  tsrc * (source_model + scale_factor * background_model)

where tsrc is the source exposure time and scale_factor corrects the
background model to the source aperture.

The existing code uses the same scaling factor as above - namely

  scale_factor = r_exposure * r_backscale * r_areascal

but this is wrong. The source and background models return rates
and not counts. This is why the source model starts with 'tsrc * '
since you need to convert to counts. This means that the ratio
between the source and background exposure times DOES NOT FACTOR
into scale_factor. It is unclear to me whether AREASCAL should be
included or not: XSPEC treats it as an exposure correction, so
I am also dropping it here, but tis *could* be wrong.

The correct scale factor should therefore be

  scale_factor = r_backscale

that is, we just correct for the aperture size when adding the
background model to the source aperture.

Fortunately this is not lilely to be a big problem with existing
users as I **IMAGINE** that the source and background exposures
are often the same, and so it doesn't matter.

It will matter for users fitting a background model to "blank-sky
datasets", where the exposure time is different in the source and
background spectra.