Tests: change BACKSCAL and background model amplitude
The ratio of BACKSCAL and exposure factors was 1/2.5 * 0.1/0.4 = 0.1
which is the same as the bin-width in energy units. This changes
the BACKSCAL ratio to 0.2/0.4 instead, to make it more obvious that
factors of 10 are from the energy grid.

The background model amlitude was changed from 0.1 to 0.4 for a
similar reason.