Tests: adjust plot tests
The aim is to make it easier to follow the terms (e.g. the expected
model values) by using names (e.g. BGND_NORM) rather than numeric
constants. The energy scale has been changed so that it's not a
constant, as the use of a 0.1 keV bin width lead to a factor of 10
in values which is too easy to confuse with other terms.