models: try to use consistent ArithmeticXXXModel handling
We can say '23 + mdl' and internally the BinaryOpModel class lifts
the constant term to an ArithmeticConstantModel, but we can not
use an ArithmeticConstantModel instead of the 23. There are also
other places where the checks for what type of object can be
passed through only allows an ArithmeticModel when it could
allow ArithmeticConstantModel or ArithmeticFunctionModel
objects as well.

There may be some long-term consequence for these changes, but it's
not obvious and there's no documentation to describe how these
classes are meant to interact.