Move response generating for source+background to background module
This takes the response generation for the case of creating the
combined response for source + background out of the Session class
and into the sherpa.astro.background module (which is where the
background model class is).

At present this doesn't buy us anything, but the logic in the
new add_response function will chnage significantly.