Tests: reduce sensitivity of test to Python version
The test, which checks the string representation of the model
expressions, is sensitive to the numbers in the input file (the
exposure time and BACKSCAL values). This can complicate the test
as the numbers display different with different Python versions,
so I explicitly set the valkues to (hopefully) more-palatable
values for Python.