pragbayes: fix an error condition
If a covariance error was raised when drawing a proposal the code
woud try to log the information and then reject the proposal.
Unfortunately the log message attempted to include the proposed
parameters, which by definition can not exist (at least if this
happens on the first loop), so I have changed the log message
to show the values of the current params (from which the draw
was attempted).

This is really hard to trigger, so there isn't a test for it.