Simplify logic slightly and add more test coverage
I added a DataErr catch case around calls to bkg.notice, but it
turns out it's not needed because the notice method itself catches
the DataErr case and displays a INFO message instead.

I have also added some basic tests to check the behavior and
warning message created by the code (and through in a background
check as well).