Move the datadir routines to a top-level package
sherpa/conftest.py has to access the get_datadir/set_datadir routines,
but this required importing sherpa.utils.testing, which caused
sherpa.utils._utils to be imported, and this was causing problems
on Travis with tests run via pytest.

One solution is to move these routines to a top-level package
which should be easier to import. As I don't understand why there's
a problem in the first place, this is all speculation.