Enforce model dimensions when combining models
We now track the dimensionality of models, using the ndim attribute
of the Model class, and enforce the constraint that model dimensions
must match when creating a CompositeModel object. If the dimensions
do not match then a ModelErr is raised.

The RegriddableModel1D and RegriddableModel2D classes set ndim to
1 and 2, respectively, which is where models are expected to pick
them up from. Note that some models, such as ArithmeticConstantModel
and ArithmeticFunctionModel, which leave ndim as None because it is
not possible to define their dimensionality from their use (e.g.
a constant, when converted into a ArithmeticConstantModel, can be
used for any model dimensionality, and for ArithmeticFunctionModel
it defines on the function that is being wrapped).

To support those cases where a dimension can not be given, the
CompositeModel check is only done for model components whose ndim
argument is not None. So combining an expression with ndim=None and
ndim=2 will create a combined model with ndim=2.

There may be some cases, such as table models, which are not handled
correctly.

I have also added a check to the ArithmeticConstantModel class to
ensure that the value used is either a scalar or a 1D array, since
the way the model evaluation works we can only use one of these
two arrays.