Add regression test for TableModel
The table model currently doesn't enforce the dimensionality of the
model, for a similar reason to ArithmeticConstantModel.

For some reason load_table_model fails when there's no I/O module
(I would have expected it could fall through to the internal I/O
code, but apparently not), so this test needs the requires_fits
decorator.