Tests: move sherpa.utils.testing to sherpa.testing
The aim of this move is to remove the need to implicitly import the
sherpa.utils._utils module, as it is causing problems on Travis.

This has meant removing the use of sao_fcmp in comparing numbers,
and relying instead on pytest.approx. We can also remove a method
in SherpaTestCase which we don't use.