MPL: change how histograms are drawn
I don't like the "pseudo" histogram style options we have when plotting
data - i.e. the drawstyle=steps* options. For example, the start/end
bins are not fully covered with the steps option.

An alternative would be to use rectangles to draw each bin, which
would be similar to the matplotlib hist command (except that we
don't need to bin the data up). I dislike this since I don't want
the bin edges to be drawn down to 0 for each bin.

The approach I use is to explicitly create a curve that draws each
bin; that is, with coordinates xlo_i, xhi_i, y_i we create the set
of points

  xlo_0, y_0
  xhi_0, y_0
  xlo_1, y_1
  xhi_1, y_1
  xlo_2, y_2
  xhi_2, y_2
  ...

and then draw straight lines between each point. For non-consecutive
bins, that is xhi_i < xlo_i+1, an extra point with value (nan, nan)
is added between them, so that the two bins are not connected.

We then draw a separate set of points for the "point/error bar"
case, which uses the mid-point of each bin.

Note that the line style is only set/changeable for the line (the
first Line2D) and the marker options are only set/changeable for the
points (the second Line2D).