FEATURE: support alpha transparency for matplotlib plots
The transparency can be set at the plot site - e.g.

    plot_fit(alpha=0.7)

or via the 'alpha' setting in the plot preferences.

Thanks to the current design of the vline and hline methods of
sherpa.plot.Plot we can not set the alpha setting for these,
so

    plot_cdf(alpha=0.4)

will only change the transparency of the data, not the lower,
upper, and median lines.