Draw Data1DInt models as a histogram and not as x,y connected points
The user interface is unchanged, but if you have a Data1DInt dataset
then

    plot_model()
    plot_model_component()
    plot_source()
    plot_source_component()
    plot_fit()
    plot_fit_resid()

will now display the model as a histogram (i.e. horizontal lines
across the full width of the bin, connected by vertical lines). It
also affects the return value of

    get_model_plot()
    get_model_component_plot()
    get_source_plot()
    get_source_component_plot()

A new model class is added to sherpa.plot to support drawing a model
using a histogram grid.

The UI layers now have to special case Data1DInt datasets when
handling a model. It is rather messy due to the centralised logic
in _prepare_plotobj, and the fact that information is duplicated
in the sherpa.astro.ui layer rather than just calling to the
superclass when needed (and the design makes this hard to see
whether we can reliably do this). This is all to note that we need
a big PR revamping the whole plotting code.